refactor: simplify draft logic
