+++ app/components/caption-editor.tsx
@@ -1,31 +1,29 @@
 import { Transition } from "@headlessui/react";
-import { tinyassert, wrapError } from "@hiogawa/utils";
-import {
-  toArraySetState,
-  useRafLoop,
-  useStableCallback,
-} from "@hiogawa/utils-react";
+import { toArraySetState, useRafLoop } from "@hiogawa/utils-react";
 import { useMutation } from "@tanstack/react-query";
 import React from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "react-hot-toast";
-import { z } from "zod";
 import { useDocumentEvent } from "../utils/hooks-client-utils";
 import { cls, zipMax } from "../utils/misc";
-import { useEffectNoStrict } from "../utils/misc-react";
 import type { CaptionEntry } from "../utils/types";
 import {
   YoutubePlayer,
   stringifyTimestamp,
   usePlayerLoader,
 } from "../utils/youtube";
+import { CaptionEditorEntry } from "./caption-editor-utils";
 import { transitionProps } from "./misc";
 import { useModal } from "./modal";
 
 // TODO: virtualize list? (no perf problem when used for short 3 min song)
 // TODO: support "half manual" mode
 
-export function CaptionEditor(props: { videoId: string }) {
+export function CaptionEditor(props: {
+  videoId: string;
+  defaultValue: CaptionEditorEntry[];
+  onChange: (v: CaptionEditorEntry[]) => void;
+}) {
   const [player, setPlayer] = React.useState<YoutubePlayer>();
 
   const { ref, isLoading } = usePlayerLoader(
@@ -33,7 +31,15 @@ export function CaptionEditor(props: { videoId: string }) {
     { onReady: setPlayer }
   );
 
-  const [entries, setEntries] = React.useState<CaptionEditorEntry[]>([]);
+  const [entries, setEntries] = React.useState<CaptionEditorEntry[]>(
+    props.defaultValue ?? []
+  );
+
+  React.useEffect(() => {
+    if (props.defaultValue !== entries) {
+      props.onChange(entries);
+    }
+  }, [entries]);
 
   const setEntriesExtra = toArraySetState(setEntries);
 
@@ -154,28 +160,6 @@ export function CaptionEditor(props: { videoId: string }) {
       player.seekTo(player.getCurrentTime() + 5);
     }
   });
-
-  //
-  // auto load/save in localstorage
-  //
-  const draft = createDraftUtils(props.videoId);
-  const draftSet = useDebounce(draft.set, 300);
-
-  // auto load on mount
-  useEffectNoStrict(() => {
-    tinyassert(entries.length === 0);
-    const found = draft.get();
-    if (found) {
-      setEntries(found);
-    }
-  }, []);
-
-  // auto save on change
-  useEffectNoStrict(() => {
-    if (entries.length > 0) {
-      draftSet(entries);
-    }
-  }, [entries]);
 
   return (
     <div className="h-full flex flex-col items-center h-full">
@@ -332,52 +316,6 @@ export function CaptionEditor(props: { videoId: string }) {
   );
 }
 
-function createDraftUtils(key: string) {
-  const draftKey = `ytsub:useDraft:${key}`;
-
-  function get(): CaptionEditorEntry[] | undefined {
-    const item = window.localStorage.getItem(draftKey);
-    if (item) {
-      const parsed = wrapError(() =>
-        Z_CAPTION_EDITOR_ENTRY.array().parse(JSON.parse(item))
-      );
-      if (parsed.ok) {
-        return parsed.value;
-      }
-      console.error("createDraftUtils", parsed.value);
-    }
-    return;
-  }
-
-  function set(data: CaptionEditorEntry[]) {
-    const item = JSON.stringify(data);
-    window.localStorage.setItem(draftKey, item);
-  }
-
-  return { get, set };
-}
-
-// TODO: utils
-function debounce<F extends (...args: any[]) => void>(f: F, ms: number): F {
-  let subscription: any;
-
-  function wrapper(this: unknown, ...args: unknown[]) {
-    if (typeof subscription !== "undefined") {
-      clearTimeout(subscription);
-    }
-    subscription = setTimeout(() => f.apply(this, args), ms);
-  }
-
-  return wrapper as any;
-}
-
-// TODO: utils-react
-function useDebounce<F extends (...args: any[]) => void>(f: F, ms: number) {
-  // TODO: pending?
-  f = useStableCallback(f);
-  return React.useCallback(debounce(f, ms), [ms]);
-}
-
 function createInitialEntries(
   text1: string,
   text2: string
@@ -450,16 +388,6 @@ function ImportModalForm(props: {
     </form>
   );
 }
-
-const Z_CAPTION_EDITOR_ENTRY = z.object({
-  begin: z.number(),
-  end: z.number(),
-  endLocked: z.boolean(),
-  text1: z.string(),
-  text2: z.string(),
-});
-
-type CaptionEditorEntry = z.infer<typeof Z_CAPTION_EDITOR_ENTRY>;
 
 function parseCaptionInput(input: string): string[] {
   return input
