+++ app/components/stories.tsx
@@ -5,6 +5,7 @@ import { useForm } from "react-hook-form";
 import { cls } from "../utils/misc";
 import { YoutubePlayer, usePlayerLoader } from "../utils/youtube";
 import { CaptionEditor } from "./caption-editor";
+import { createDraftUtils } from "./caption-editor-utils";
 import { SelectWrapper, VideoComponent, transitionProps } from "./misc";
 import { PopoverSimple } from "./popover";
 import {
@@ -328,6 +329,17 @@ export function TestYoutubePlayer() {
 }
 
 export function TestCaptionEditor() {
+  const videoId = "UY3N52CrTPE";
+
+  const draftUtils = createDraftUtils(videoId);
+  const draftData = React.useMemo(() => draftUtils.get() ?? [], []);
+
   // https://www.youtube.com/watch?v=UY3N52CrTPE
-  return <CaptionEditor videoId="UY3N52CrTPE" />;
+  return (
+    <CaptionEditor
+      videoId="UY3N52CrTPE"
+      defaultValue={draftData}
+      onChange={(data) => draftUtils.set(data)}
+    />
+  );
 }
