+++ app/routes/videos/caption-editor.tsx
@@ -1,5 +1,7 @@
 import { tinyassert } from "@hiogawa/utils";
+import React from "react";
 import { CaptionEditor } from "../../components/caption-editor";
+import { createDraftUtils } from "../../components/caption-editor-utils";
 import { ROUTE_DEF } from "../../misc/routes";
 import { useLoaderDataExtra } from "../../utils/loader-utils";
 import { makeLoader } from "../../utils/loader-utils.server";
@@ -32,11 +34,16 @@ export const loader = makeLoader(async ({ ctx }) => {
 export default function Page() {
   const { videoId } = useLoaderDataExtra() as LoaderData;
 
-  // TODO: move draft loading logic to outside of CaptionEditor?
+  const draftUtils = createDraftUtils(videoId);
+  const draftData = React.useMemo(() => draftUtils.get() ?? [], []);
 
   return (
     <div className="p-2 h-full">
-      <CaptionEditor videoId={videoId} />
+      <CaptionEditor
+        videoId={videoId}
+        defaultValue={draftData}
+        onChange={(data) => draftUtils.set(data)}
+      />
     </div>
   );
 }
